Add option to set the temporary directory path on Windows.
As of May 2020, MiKTeX does no longer support short filenames, so if the path to the temporary directory contains short filenames, LaTeX compilation will fail.

In the current framework it seems to be impossible to obtain the full path with only long filenames easily, so present the user with an easy way to change it if necessary.

Also display the "auto-detected" path on the firstrun page.

On OSs other than Windows this option is unset, unused, and hidden in the options and firstrun pages.